\name{00.tsfa.Intro}
\alias{00.tsfa.Intro}
\docType{package}

\title{Time Series Factor Analysis (TSFA)}

\description{TSFA extends standard factor analysis (FA) to 
time series data. Rotations methods can be applied as in FA.
A dynamic model of the factors is not assumed, but could be 
estimated separately using the extracted factors.
}

\details{See \code{\link{tsfa-package}} ( in the help system use
   package?tsfa or ?"tsfa-package") for an overview.
}
\keyword{package}

