\name{factors}
\alias{factors}
\alias{factors.TSFmodel}
\alias{factors.EstEval}

\title{Extract Time Series Factors from an Object}
\usage{
    factors(x)
    \method{factors}{TSFmodel}(x)
    \method{factors}{EstEval}(x)
}
\arguments{
    \item{x}{an object.}
}
\description{Extract time series factors from an object.}

\value{factor series.}

\seealso{
   \code{\link{TSFmodel}},
   \code{\link{estTSF.ML}},
   \code{\link{simulate.TSFmodel}}
}
\author{Paul Gilbert}
\concept{TSFA}
\keyword{ts}
